+++ test/com/mobiletheatertech/plot/DrapeBaseTest.java
@@ -50,7 +50,7 @@ public class DrapeBaseTest {
 
         assertEquals( TestHelpers.accessDouble(base, "x"), x );
         assertEquals( TestHelpers.accessDouble(base, "y"), y );
-        assertEquals( TestHelpers.accessDouble(base, "z"), 0.0 );
+        assertEquals( TestHelpers.accessDouble(base, "z"), (Double)0.0 );
     }
 
     @Test
